I think my points are printed in the right place

+++ inc/struct.h
@@ -6,7 +6,7 @@
 /*   By: momrane <momrane@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/13 12:17:51 by momrane           #+#    #+#             */
-/*   Updated: 2024/01/18 17:05:44 by momrane          ###   ########.fr       */
+/*   Updated: 2024/01/18 20:31:07 by momrane          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -52,7 +52,7 @@ typedef struct s_env
 	int					angle;
 	int					zoom;
 	int					scale;
-	int					space;
+	float				space;
 	struct s_pt			origin;
 	struct s_parsing	data;
 	struct s_img		img;
